[use] Remove hb-ot-shape-complex-use.hh
Inline into ragel machine.
