+++ src/Makefile.sources
@@ -125,7 +125,6 @@ HB_BASE_sources = \
 	hb-ot-shape-complex-thai.cc \
 	hb-ot-shape-complex-use-table.cc \
 	hb-ot-shape-complex-use.cc \
-	hb-ot-shape-complex-use.hh \
 	hb-ot-shape-complex-vowel-constraints.cc \
 	hb-ot-shape-complex-vowel-constraints.hh \
 	hb-ot-shape-complex.hh \
